Move GetOrderReference to base payment provider
