+++ src/Vendr.Contrib.PaymentProviders.Reepay/ReepayCheckoutPaymentProvider.cs
@@ -35,39 +35,6 @@ namespace Vendr.Contrib.PaymentProviders.Reepay
             new TransactionMetaDataDefinition("reepayCustomerHandle", "Reepay Customer Handle")
         };
 
-        public override OrderReference GetOrderReference(HttpRequestBase request, ReepayCheckoutSettings settings)
-        {
-            try
-            {
-                var reepayEvent = GetReepayWebhookEvent(request, settings);
-                if (reepayEvent != null)
-                {
-                    if (!string.IsNullOrWhiteSpace(reepayEvent.Invoice) && 
-                        (reepayEvent.EventType == WebhookEventType.InvoiceAuthorized ||
-                        reepayEvent.EventType == WebhookEventType.InvoiceSettled ||
-                        reepayEvent.EventType == WebhookEventType.SubscriptionCreated))
-                    {
-                        var clientConfig = GetReepayClientConfig(settings);
-                        var client = new ReepayClient(clientConfig);
-                        var metadata = client.GetInvoiceMetaData(reepayEvent.Invoice);
-                        if (metadata != null)
-                        {
-                            if (metadata.TryGetValue("orderReference", out object orderReference))
-                            {
-                                return OrderReference.Parse(orderReference.ToString());
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Vendr.Log.Error<ReepayCheckoutPaymentProvider>(ex, "Reepay - GetOrderReference");
-            }
-
-            return base.GetOrderReference(request, settings);
-        }
-
         public override PaymentFormResult GenerateForm(OrderReadOnly order, string continueUrl, string cancelUrl, string callbackUrl, ReepayCheckoutSettings settings)
         {
             var currency = Vendr.Services.CurrencyService.GetCurrency(order.CurrencyId);
